data-ingestion-kafka/requirements.txt
requests
kafka-python==2.0.2
